+++ util/StorageConfig.hpp
@@ -8,7 +8,7 @@ namespace store {
 struct StorageConfig {
     uint16_t gcIntervall = 60;
     size_t totalMemory = TOTAL_MEMORY;
-    int numScanThreads = 2;
+    size_t numScanThreads = 2;
     size_t hashMapCapacity = HASHMAP_CAPACITY;
 };
 } // namespace store
